true clock + time gems

+++ common/src/main/java/io/github/chromonym/chronoception/mixin/ClientWorldMixin.java
@@ -3,9 +3,6 @@ package io.github.chromonym.chronoception.mixin;
 import java.util.function.Supplier;
 
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import io.github.chromonym.chronoception.client.ChronoceptionClient;
 import net.minecraft.client.world.ClientWorld;
@@ -29,11 +26,11 @@ public abstract class ClientWorldMixin extends World {
 
     @Override
     public long getTimeOfDay() {
-        return super.getTimeOfDay() + ChronoceptionClient.playerData.offset;
+        return super.getTimeOfDay() + (long)ChronoceptionClient.playerData.offset;
     }
 
     @Override
     public long getLunarTime() {
-        return super.getLunarTime() + ChronoceptionClient.playerData.offset;
+        return super.getLunarTime() + (long)ChronoceptionClient.playerData.offset;
     }
 }
